Extracted the table part of the JiraUserIssuesViewCard into a separate component JiraUserIssuesTable
