+++ .changeset/five-seas-itch.md
@@ -1,0 +1,5 @@
+---
+'@axis-backstage/plugin-jira-dashboard': patch
+---
+
+Extracted the table component from JiraUserIssuesViewCard into an exported component JiraUserIssuesTable, and added styling options to the table (or its outer Card) to adapt it more freely
