+++ plugins/jira-dashboard/src/components/JiraUserIssuesTable/index.ts
@@ -1,0 +1,5 @@
+export type {
+  TableComponentProps,
+  JiraUserIssuesTableProps,
+} from './JiraUserIssuesTable';
+export { JiraUserIssuesTable } from './JiraUserIssuesTable';
